refactor: extract draw function into lib
