+++ frontend/components/MapImage.tsx
@@ -5,195 +5,12 @@ import useTileGroup from "@/hooks/useTileGroup";
 import useTilesetData from "@/hooks/useTilesetData";
 import { useRawMapStore } from "@/store/mapStore";
 import { useImages } from "@/hooks/useImage";
-import { Terrain } from "@/types/schemas/Terrain";
-import { Unit } from "@/types/schemas/Unit";
-import { Flingy } from "@/types/schemas/Flingy";
-import { Sprite } from "@/types/schemas/Sprite";
-import { SCImageBundle } from "@/types/SCImage";
-import { createTeamColorUnitImage, fetchFrameImage } from "@/lib/scimage";
-import { Location } from "@/types/schemas/Location";
-
-const TILE_SIZE = 32;
-
-function drawMegatile(
-  ctx: OffscreenCanvasRenderingContext2D,
-  x: number,
-  y: number,
-  rgbData: Uint8Array,
-) {
-  const imageData = ctx.createImageData(TILE_SIZE, TILE_SIZE);
-
-  for (let i = 0; i < TILE_SIZE * TILE_SIZE; i++) {
-    const r = rgbData[i * 3];
-    const g = rgbData[i * 3 + 1];
-    const b = rgbData[i * 3 + 2];
-
-    imageData.data[i * 4] = r;
-    imageData.data[i * 4 + 1] = g;
-    imageData.data[i * 4 + 2] = b;
-    imageData.data[i * 4 + 3] = 255; // Alpha
-  }
-
-  ctx.putImageData(imageData, x * TILE_SIZE, y * TILE_SIZE);
-}
-
-function getTerrainImage(
-  terrain: Terrain,
-  tileGroup: number[][],
-  tilesetData: Uint8Array,
-) {
-  const terrainCanvas = new OffscreenCanvas(
-    terrain.size.width * TILE_SIZE,
-    terrain.size.height * TILE_SIZE,
-  );
-  const terrainCtx = terrainCanvas.getContext("2d")!;
-
-  for (let y = 0; y < terrain.size.height; y++) {
-    for (let x = 0; x < terrain.size.width; x++) {
-      const tile = terrain.tile_id[y][x];
-      const megatileID = tileGroup[tile.group][tile.id];
-      const offset = megatileID * 3072;
-      const rgbData = tilesetData.slice(offset, offset + 3072);
-      drawMegatile(terrainCtx, x, y, rgbData);
-    }
-  }
-
-  return terrainCanvas.transferToImageBitmap();
-}
-
-async function getPlacedUnitImage(
-  terrain: Terrain,
-  placedUnit: Unit[],
-  flingy: Flingy[],
-  sprite: Sprite[],
-  SCImages: Map<number, SCImageBundle>,
-): Promise<ImageBitmap> {
-  const canvas = new OffscreenCanvas(
-    terrain.size.width * TILE_SIZE,
-    terrain.size.height * TILE_SIZE,
-  );
-  const ctx = canvas.getContext("2d")!;
-
-  const tasks = placedUnit.map(async (unit) => {
-    const flingyID = unit.unit_definition.specification.graphics;
-    const spriteID = flingy[flingyID].sprite;
-    const imageID = sprite[spriteID].image;
-    const image = SCImages.get(imageID);
-
-    if (!image || !image.diffuse) return;
-
-    if (image.teamColor) {
-      const playerColor = unit.owner ? unit.owner.rgb_color : [255, 255, 0];
-      const diffuse = await fetchFrameImage({
-        image: image.diffuse,
-        frame: 0,
-        meta: image.meta,
-      });
-      const teamColor = await fetchFrameImage({
-        image: image.teamColor,
-        frame: 0,
-        meta: image.meta,
-      });
-      const colored = createTeamColorUnitImage(
-        diffuse,
-        teamColor,
-        playerColor as [number, number, number],
-      );
-
-      ctx.drawImage(
-        colored,
-        unit.transform!.position.x,
-        unit.transform!.position.y,
-      );
-    } else {
-      ctx.drawImage(
-        await fetchFrameImage({
-          image: image.diffuse,
-          frame: 0,
-          meta: image.meta,
-        }),
-        unit.transform!.position.x,
-        unit.transform!.position.y,
-      );
-    }
-  });
-
-  await Promise.all(tasks);
-
-  return canvas.transferToImageBitmap();
-}
-
-async function getPlacedSpriteImages(
-  terrain: Terrain,
-  placedSprite: Sprite[],
-  SCImages: Map<number, SCImageBundle>,
-) {
-  const canvas = new OffscreenCanvas(
-    terrain.size.width * TILE_SIZE,
-    terrain.size.height * TILE_SIZE,
-  );
-  const ctx = canvas.getContext("2d")!;
-
-  const tasks = placedSprite.map(async (sprite) => {
-    const imageID = sprite.image;
-    const image = SCImages.get(imageID);
-
-    if (!image) return;
-    ctx.drawImage(
-      await fetchFrameImage({
-        image: image.diffuse,
-        frame: 0,
-        meta: image.meta,
-      }),
-      sprite.transform!.position.x,
-      sprite.transform!.position.y,
-    );
-  });
-
-  await Promise.all(tasks);
-
-  return canvas.transferToImageBitmap();
-}
-
-function getLocationImage(terrain: Terrain, locations: Location[]) {
-  const thickness = 3;
-  const canvas = new OffscreenCanvas(
-    terrain.size.width * TILE_SIZE,
-    terrain.size.height * TILE_SIZE,
-  );
-  const ctx = canvas.getContext("2d")!;
-  console.log(locations);
-
-  locations.forEach((location) => {
-    if (location.id != 63) {
-      const width = location.position.right - location.position.left;
-      const height = location.position.bottom - location.position.top;
-      ctx.fillStyle = "rgba(0, 0, 255, 0.15)";
-      ctx.fillRect(
-        location.position.left,
-        location.position.top,
-        width,
-        height,
-      );
-      ctx.fillStyle = "rgba(0, 0, 0, 0.15)";
-      ctx.strokeRect(
-        location.position.left,
-        location.position.top,
-        width + thickness,
-        height + thickness,
-      );
-      ctx.font = "40px serif";
-      ctx.fillStyle = "rgb(255, 255, 255)";
-      ctx.fillText(
-        location.name,
-        location.position.left,
-        location.position.top,
-      );
-    }
-  });
-
-  return canvas.transferToImageBitmap();
-}
+import {
+  getLocationImage,
+  getPlacedSpriteImages,
+  getPlacedUnitImage,
+} from "@/lib/scimage";
+import { TILE_SIZE, getTerrainImage } from "@/lib/scterrain";
 
 export const MapImage = () => {
   const viewportCanvasRef = useRef<HTMLCanvasElement | null>(null);
