+++ frontend/lib/scimage.ts
@@ -1,4 +1,10 @@
-import { FrameMeta } from "@/types/SCImage";
+import { Flingy } from "@/types/schemas/Flingy";
+import { Sprite } from "@/types/schemas/Sprite";
+import { Terrain } from "@/types/schemas/Terrain";
+import { Unit } from "@/types/schemas/Unit";
+import { FrameMeta, SCImageBundle } from "@/types/SCImage";
+import { Location } from "@/types/schemas/Location";
+import { TILE_SIZE } from "./scterrain";
 
 type props = { image: Blob; frame: number; meta: FrameMeta };
 export async function fetchFrameImage({ image, frame, meta }: props) {
@@ -52,4 +58,138 @@ export function createTeamColorUnitImage(
   }
 
   return unitCanvas;
+}
+
+export async function getPlacedUnitImage(
+  terrain: Terrain,
+  placedUnit: Unit[],
+  flingy: Flingy[],
+  sprite: Sprite[],
+  SCImages: Map<number, SCImageBundle>,
+): Promise<ImageBitmap> {
+  const canvas = new OffscreenCanvas(
+    terrain.size.width * TILE_SIZE,
+    terrain.size.height * TILE_SIZE,
+  );
+  const ctx = canvas.getContext("2d")!;
+
+  const tasks = placedUnit.map(async (unit) => {
+    const flingyID = unit.unit_definition.specification.graphics;
+    const spriteID = flingy[flingyID].sprite;
+    const imageID = sprite[spriteID].image;
+    const image = SCImages.get(imageID);
+
+    if (!image || !image.diffuse) return;
+
+    if (image.teamColor) {
+      const playerColor = unit.owner ? unit.owner.rgb_color : [255, 255, 0];
+      const diffuse = await fetchFrameImage({
+        image: image.diffuse,
+        frame: 0,
+        meta: image.meta,
+      });
+      const teamColor = await fetchFrameImage({
+        image: image.teamColor,
+        frame: 0,
+        meta: image.meta,
+      });
+      const colored = createTeamColorUnitImage(
+        diffuse,
+        teamColor,
+        playerColor as [number, number, number],
+      );
+
+      ctx.drawImage(
+        colored,
+        unit.transform!.position.x,
+        unit.transform!.position.y,
+      );
+    } else {
+      ctx.drawImage(
+        await fetchFrameImage({
+          image: image.diffuse,
+          frame: 0,
+          meta: image.meta,
+        }),
+        unit.transform!.position.x,
+        unit.transform!.position.y,
+      );
+    }
+  });
+
+  await Promise.all(tasks);
+
+  return canvas.transferToImageBitmap();
+}
+
+export async function getPlacedSpriteImages(
+  terrain: Terrain,
+  placedSprite: Sprite[],
+  SCImages: Map<number, SCImageBundle>,
+) {
+  const canvas = new OffscreenCanvas(
+    terrain.size.width * TILE_SIZE,
+    terrain.size.height * TILE_SIZE,
+  );
+  const ctx = canvas.getContext("2d")!;
+
+  const tasks = placedSprite.map(async (sprite) => {
+    const imageID = sprite.image;
+    const image = SCImages.get(imageID);
+
+    if (!image) return;
+    ctx.drawImage(
+      await fetchFrameImage({
+        image: image.diffuse,
+        frame: 0,
+        meta: image.meta,
+      }),
+      sprite.transform!.position.x,
+      sprite.transform!.position.y,
+    );
+  });
+
+  await Promise.all(tasks);
+
+  return canvas.transferToImageBitmap();
+}
+
+export function getLocationImage(terrain: Terrain, locations: Location[]) {
+  const thickness = 3;
+  const canvas = new OffscreenCanvas(
+    terrain.size.width * TILE_SIZE,
+    terrain.size.height * TILE_SIZE,
+  );
+  const ctx = canvas.getContext("2d")!;
+  console.log(locations);
+
+  locations.forEach((location) => {
+    if (location.id != 63) {
+      const width = location.position.right - location.position.left;
+      const height = location.position.bottom - location.position.top;
+      ctx.fillStyle = "rgba(0, 0, 255, 0.15)";
+      ctx.fillRect(
+        location.position.left,
+        location.position.top,
+        width,
+        height,
+      );
+      ctx.fillStyle = "rgba(0, 0, 0, 0.15)";
+      ctx.strokeRect(
+        location.position.left,
+        location.position.top,
+        width + thickness,
+        height + thickness,
+      );
+      ctx.font = "40px serif";
+      ctx.fillStyle = "rgb(255, 255, 255)";
+      ctx.fillText(
+        location.name,
+        location.position.left,
+        location.position.top,
+      );
+    }
+  });
+
+  return canvas.transferToImageBitmap();
 }
